tpms-backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.0
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.0
celery==5.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-telegram-bot==20.7
cryptography==41.0.0
httpx==0.26.0
aioredis==2.0.1
asyncpg==0.29.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
black==23.12.0
flake8==7.0.0
mypy==1.8.0
prometheus-client==0.19.0
structlog==23.2.0
python-dotenv==1.0.0